refactor: 비슷한 interface 분리
