+++ src/interface/DropdownInterfaces.ts
@@ -1,0 +1,15 @@
+interface DropdownOption {
+  value: string;
+  content: string;
+}
+
+interface DropdownProps {
+  options: DropdownOption[];
+  label?: string;
+  name?: string;
+  id?: string;
+  className?: string;
+  isRequired: boolean;
+}
+
+export { DropdownOption, DropdownProps };
